+++ src/main/resources/static/createTable.sql
@@ -2,7 +2,7 @@
 DROP TABLE IF EXISTS PRODUCT;
 create table PRODUCT
 (
-id integer PRIMARY KEY,
+id BIGINT PRIMARY KEY,
 title varchar(255),
 description varchar(255) not null,
 price integer not null,
@@ -11,7 +11,7 @@ city varchar(255) not null,
 
 create table USER
 (
-id integer not null,
+id BIGINT not null,
 login varchar(255) not null,
 password varchar(255) not null,
 primary key (id)
@@ -19,15 +19,15 @@ primary key (id)
 
 
 create table ROLESOFUSERS (
-id integer not null,
+id BIGINT not null,
 role_name varchar(50) not null,
 primary key (id)
 );
 
 create table USER_ROLES
 (
-user_id integer not null,
-userrole_id integer not null
+user_id BIGINT not null,
+userrole_id BIGINT not null
 );
 
 INSERT INTO PRODUCT(id, title, description, price, city)
